fix: Renamed RhinoConverterToNative to ToHost
