+++ DUI3-DX/Converters/Rhino/Speckle.Converters.Rhino7/ToHost/RhinoConverterToHost.cs
@@ -1,9 +1,9 @@
 ﻿using Speckle.Converters.Common;
 using Speckle.Core.Models;
 
-namespace Speckle.Converters.Rhino7;
+namespace Speckle.Converters.Rhino7.ToHost;
 
-public class RhinoConverterToNative : ISpeckleConverterToHost
+public class RhinoConverterToHost : ISpeckleConverterToHost
 {
   public object Convert(Base target) => throw new NotImplementedException();
 }
